requirements.txt
numpy>=1.21.0
matplotlib>=3.5.0
scipy>=1.7.0
opencv-python>=4.5.0
Pillow>=8.0.0
Flask>=2.0.0

# Optional dependencies for enhanced visualization
# seaborn>=0.11.0
# plotly>=5.0.0
